zad2/Naziemiec_Blazej_2/pyproject.toml
[project]
name = "naziemiec-blazej-2"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "slowapi>=0.1.9",
    "uvicorn>=0.42.0",
]
